backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
